[PSL] Add testing framework & remove useless files.
